refactor: move tensor_kind.h + storage_tier.h to src/core/
Both headers were in include/imp/ but they're not part of the public
C API — they use namespace imp { ... } and enum class, neither of
which compiles in C. CONTRIBUTING.md / CLAUDE.md state that
include/imp/ holds the C-compatible public surface and src/ holds
internal types.

Moved to src/core/. Updated 9 includes (3 in src/graph and src/runtime
and src/model, 1 in src/core/tensor.h, 1 in src/model/tensor_kind_name.cpp,
3 in tests/, 1 in src/model/tensor_kind_table.h).

include/imp/ now only contains the four public headers:
imp.h, config.h, error.h, types.h.

Verified: 37/37 unit tests pass after move.
