+++ src/model/tensor_kind_name.cpp
@@ -1,4 +1,4 @@
-#include "imp/tensor_kind.h"
+#include "core/tensor_kind.h"
 
 namespace imp {
